% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hhdat}
\alias{hhdat}
\title{PSRC Household-Level Data}
\format{
A data frame with 97 variables
}
\source{
\url{https://www.psrc.org/household-travel-survey-program}
}
\usage{
hhdat
}
\description{
Household-level dataset from the 2017-2019 Pacific Southwest Regional Council Household Travel Survey Public Release 2.
}
\keyword{datasets}
