% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prdat}
\alias{prdat}
\title{PSRC Person-Level Data}
\format{
A data frame with 124 variables
}
\source{
\url{https://www.psrc.org/household-travel-survey-program}
}
\usage{
prdat
}
\description{
Person-level dataset from the 2017-2019 Pacific Southwest Regional Council Household Travel Survey Public Release 2.
}
\keyword{datasets}
