% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trdat}
\alias{trdat}
\title{PSRC Travel Diary Data}
\format{
A data frame with 122 variables
}
\source{
\url{https://www.psrc.org/household-travel-survey-program}
}
\usage{
trdat
}
\description{
Travel diary dataset from the 2017-2019 Pacific Southwest Regional Council Household Travel Survey Public Release 2.
}
\keyword{datasets}
